+++ src/api/ingredient.ts
@@ -1,0 +1,4 @@
+import axiosInstance from ".";
+export const reqIngredientFindByProduct = (productId: number) => {
+  return axiosInstance.post("/ingredient/findByProduct", { productId });
+};
