+++ src/api/routineItem.ts
@@ -1,0 +1,5 @@
+import axiosInstance from ".";
+
+export const reqRoutineItemFindByRoutine = (routineId: number) => {
+  return axiosInstance.post("/routineItem/findByRoutine", { routineId });
+};
