+++ src/types/user.ts
@@ -7,14 +7,7 @@ export type IUser = {
   email?: string;
   gender: string;
 };
-// export interface IUserUpdateOptions {
-//   name?: string;
-//   password?: string;
-//   avatar?: string;
-//   phone?: string;
-//   email?: string;
-//   gender?: string;
-// }
+
 export type IUserUpdateOptions = Pick<
   IUser,
   "name" | "password" | "avatar" | "phone" | "email" | "gender"
